Moving Swift class signature transformation extractor to a separate file and improving structure of ValueTransformer to plan for better support debugging
